attempt to test consul config provider in a configuration
I am chasing some odd behavior in my app where using
the reload feature triggers a change event properly
but seems to return a different result when I have
additional providers configured after the consul provider.

this commit adds a few test cases attempting to extend
the exisiting ConsulConfigurationProviderTests+ReloadOnChange
tests to add the _source into a ConfigurationBuilder and
attempt to use it that way.

unfortunately building the ConfigurationRoot fails
with the required 'Test' configuration missing, presumably
because of the way that the source data is mocked being
insufficient to provide the configuration builder
what it needs.